v0.12.5

+++ scripts/delete_records.py
@@ -12,18 +12,22 @@ from sqlalchemy import not_, type_coerce
 
 from invenio_db import db
 from invenio_files_rest.models import Bucket, ObjectVersion
-from invenio_pidstore.models import PersistentIdentifier
+from invenio_pidstore.models import PersistentIdentifier, RecordIdentifier
 from invenio_records_files.models import RecordsBuckets
 from invenio_records.models import RecordMetadata
 
 
 def _delete_record(record):
-    rec_bucket = RecordsBuckets.query.filter(RecordsBuckets.record == record).one()
-    bucket = rec_bucket.bucket
-    ObjectVersion.query.filter(ObjectVersion.bucket == bucket).delete()
-    db.session.delete(bucket)
-    db.session.delete(rec_bucket)
-    PersistentIdentifier.query.filter(PersistentIdentifier.object_uuid == record.id).delete()
+    rec_bucket = RecordsBuckets.query.filter(RecordsBuckets.record == record).one_or_none()
+    if rec_bucket:
+        bucket = rec_bucket.bucket
+        ObjectVersion.query.filter(ObjectVersion.bucket == bucket).delete()
+        db.session.delete(bucket)
+        db.session.delete(rec_bucket)
+    pid = PersistentIdentifier.query.filter(PersistentIdentifier.object_uuid == record.id).one()
+    recid = RecordIdentifier.query.get(pid.pid_value)
+    db.session.delete(recid)
+    db.session.delete(pid)
     db.session.delete(record)
     db.session.commit()
